likes and updating and deleting user based on session implemented

+++ server/routers/postRouter.js
@@ -1,5 +1,6 @@
 const Post = require("../models/Post");
 const express = require("express");
+const { json } = require("express");
 const postRouter = express.Router();
 
 // GET all twats
@@ -9,9 +10,9 @@ postRouter.get("/posts", async (req, res) => {
 });
 
 // POST new twat
-postRouter.post("/posts", async (req, res) => {
+postRouter.post("/posts", secureWithRole("plebian"), async (req, res) => {
   const post = new Post({
-    author: req.body.author,
+    author: req.session.userName,
     content: req.body.content,
     likes: req.body.likes,
     date: req.body.date,
@@ -26,26 +27,69 @@ postRouter.get("/posts/:id", async (req, res) => {
   res.send(post);
 });
 
-//delete character from ID
-postRouter.delete("/posts/:id", async (req, res) => {
-  const post = await Post.findOneAndDelete({ _id: req.params.id });
-  res.send(post);
+//delete post from ID
+postRouter.delete("/posts/:id", secureWithRole("plebian"), async (req, res) => {
+  const twatToDelete = await Post.findOne({ _id: req.params.id });
+  if(twatToDelete.author === req.session.userName){
+    const post = await Post.findOneAndDelete({ _id: req.params.id });
+    res.status(200).json('delet')
+    res.send(post);
+  } else { 
+    res.status(401).json('You do not have the necessary priviliges')
+  }
 });
 
+//Update like 
+
 //update post from ID 
-postRouter.put("/posts/:id", async (req, res) => {
+postRouter.put("/posts/:id", secureWithRole("plebian"), async (req, res) => {
+  const twatToUpdate = await Post.findOne({ _id: req.params.id });
+  if(twatToUpdate.author === req.session.userName){
+    const post = await Post.findOneAndUpdate(
+      { _id: req.params.id },
+      { $set: { content: req.body.content } },
+      { new: true },
+      (err, doc) => {
+        if (err) {
+          console.log("Opsiiii something went wrooong");
+        }
+        console.log(doc);
+      }
+    );
+    res.status(200).json('updated')
+    res.send(post);
+  } else {
+    res.status(401).json('You do not have the necessary priviliges')
+  }
+});
+
+postRouter.post("/posts/:id", async (req, res) => {
   const post = await Post.findOneAndUpdate(
     { _id: req.params.id },
-    { $set: { content: req.body.content } },
-    { new: true },
-    (err, doc) => {
-      if (err) {
-        console.log("Opsiiii something went wrooong");
-      }
-      console.log(doc);
-    }
+    { $inc: { likes: 1 } },
+    { new: true }
   );
-  res.send(post);
+  res.status(200).json('Likes updated')
+  res.send(post)
 });
+
+//Middleware functions
+function secure(req,res,next){
+    if(req.session.userName){
+        next()
+    } else {
+        res.status(401).json('loginfirst pls')
+    }
+}
+
+function secureWithRole(role) {
+    return [secure, (req,res,next) => {
+        if(req.session.role === role){
+            next()
+        } else {
+            res.status(403).json('Check your priviliges!')
+        }
+    }]
+}
 
 module.exports = postRouter;
